Prevent including hidden from catalog courses in the `course-discovery` store or marketing frontend. EW-320
Since we have so many course duplicates (e.g. 400+ courses with several repeats of existing curriculum titles for different term runs) we need to exclude the duplicate courses from being sent to the marketing frontend should a user (e.g. `lms_catalog_service_user`) that pulls from the LMS Course/Block REST API have Staff access that would allow the payload of those services to show all course content.

We can rethink about this configuration later should we need to include these additional courses but for now it allows more control in the amount of courses that need processed to marketing frontend. Even if these courses were processed we have a `hidden` field that indicates to show or not.