Fix issue where content was not being published to marketing frontend when Model status was set to a previous Unpublished state. EW-320
The error occurred when the waffle switch `publish_course_runs_to_marketing_site` was disabled, then enabled and the `refresh_course_metadata` was rerun. This update checks to see if existing obj Model being updated has `wordpress_post_id` or not. If it does not it publishes to marketing frontend.